Allow simulation wallets to be created from standard input
It was not previously possible to create insecure or simulation
wallets with scripted input due to the terminal.ReadPassword call
failing when standard input was not a tty.  As a workaround, when the
flags or config specify the private passphrase (--pass), a new prompt
is added whether to use this value and skip the password prompt
entirely.

This also fixes a bug with the input buffering when reading the seed
could otherwise cause the wallet to loop forever from valid creation
input.

The following script provides an example creating a temporary
simulation wallet with a hardcoded private passphrase and seed:


[ -d /tmp/dcrwallet ] || mkdir /tmp/dcrwallet
dir=$(mktemp -d -p /tmp/dcrwallet)
dcrwallet --simnet --create --appdata="$dir" --pass=password <<EOF
y
n
y
8dce73948a6d51fd56530a0ee12633ef53df3add4bdd853b1ec7e3046b72934c
EOF