Remove requirement on golang.org/x/xerrors
The xerrors package was used by dcrwallet's errors package to
implement error wrapping for older versions of Go (1.12 and earlier)
which did not support these features in the stdlib.  Since these Go
releases are unsupported by the project, the wrappers can be switched
to calling the stdlib instead.

With this change, go vet was able to report one invalid format string
that was not noticed before.