Allow unpublished transactions to be saved to DB
Unpublished transactions are unmined transactions which are
specifically recorded as unpublished to the network. These
transactions are not rebroadcast on startup. It is necessary to save
unpublished transactions to prevent some previous wallet outputs from
being double spent across application restarts, or to persist
partially signed transactions.

This commit adds the database handling for recording unpublished
transactions but does not yet record any transactions in this state.
This will be used by vspd clients in the future to record unpublished
fee and ticket purchase transactions.