Move PublishTransaction high fee check to caller
We wish to add a JSON-RPC method which calls the wallet's
PublishTransaction method, but the enabling or disabling of high fee
checks works differently between the JSON-RPC and gRPC servers.  In
the JSON-RPC server, the method has its own default value for this
setting rather than using the wallet application config, while the
gRPC server uses the application settings.  One possible way to deal
with this would be to add another bool parameter for the setting to
the PublishTransaction signature.  This change instead keeps the
function signature the same, but requires callers to perform high fee
checks (if any), and gives more flexibility as to how this check is
enabled.