Allow accounts to be individually encrypted and unlocked
This commit introduces three new RPCs (setaccountpassphrase,
unlockaccount, and lockaccount) which can be used to set a unique
passphrase for and allow individual locking of particular accounts,
separate from the global keys used to encrypt all other accounts.

Argon2id is used as the KDF for individually encrypted accounts, and
XChaCha20-Poly1305 is used for authenticated encryption of account
xprivs.  The new KDF parameters use the same memory difficulty (256M)
as the scrypt parameters used for all other accounts.