Remove paragraph about public data encryption
Encrypting public data in the wallet database is a non-default option,
and generally discouraged due to all of the other public data which is
not encrypted.  Despite what the README said, there are currently no
plans to encrypt the rest of this data, as it would break some of the
ways cursors are used to iterate over saved values.