Fix some strange vsp package bugs
Unexport the Queue method, so that external callers are less likely to
abuse it.  It will panic if the queue channel has already closed.

Use returns instead of breaks when inside of a select statement and
the goal was to break out of an outer loop.  The break statement will
break out of the switch statement, not the loop.

Do not queue items from Process.  This was needlessly recursive
(Process -> queue -> Process) and caused the first Process call to
return the original error before attempting the second process
(despite waiting for 2 minutes in the foreground), by which time the
second Process call could fail due to the wallet locking after the
first Process call returned.

Remove a log of expected context cancellation errors.