vsp: Remove code duplication
This commit adds a client type which handles both GET and POST
requests for the vspd API.  The server signature is always checked,
and during POSTs, the request body is signed using the private key of
some wallet address.

This also fixes the handling of HTTP 400 status codes for bad client
requests, and will unmarshal these into a BadRequestError type.