Check for sufficiently-synced height during p2p connect
This fixes a layering issue where peers were always disconnected by
the SPV syncer if they did not advertise a sufficiently-high last
block height, when this operation is better performed by the peer code
which integrates with the address manager.  Erroring in
p2p.ConnectOutbound results in the addresses not being marked as
"good" in the address manager prior to immediately disconnecting from
the peer for not being synced.