Replace txToOutputs method, use struct for many parameters
This removes the wallet's unexported txToOutputs method by replacing
it with an authorTx method that instead of taking a dozen parameters,
uses a struct.  While not only cleaner, this also allows more fields
of the struct to be set by the method, which will be used in upcoming
commits to change where and how the created transactions are added to
the wallet and published to the network.