Use extra split output rather than selecting random UTXOs
After we create an extra split output to deal with UTXO contention in
a VSP ticket purchase, there is no need to perform UTXO selection
again to reserve an output for a single VSP fee payment.  Add an
unexported field to the purchase ticket parameters so that upon
reentry, this output can be used directly.