Implement nofilelogging config flag
This works the same as dcrd's flag by the same name.  When enabled,
logs will only be written to standard output, and not to any log files
managed by dcrwallet itself.

Of course, persistent logs are useful, so any production setup using
this flag is expected to capture standard output and redirect it to
syslogd, the systemd journal, etc.