Improve heuristic for creating additional split tx
When the additional split transaction must be created, the total
balance from the account that had been reserved for all fee txs and
all ticket purchases can be summed together, to determine if it is
enough to create the additional split transaction.  Previously, only
the partial balance of some of the selected inputs were being used,
and this had been triggering an unnecessary insufficient balance error
when attempting to create the additional split transaction would have
been a better idea.