multi: migrate to the new stdscript package
This migrates to the tagged txscript/v4 module, which uses script types
and script parsing from the txscript/v4/stdscript package.

This also updates to the following tagged dcrd modules;
- addrmgr/v2.0.0
- blockchain/stake/v4.0.0
- blockchain/standalone/v2.0.0
- chaincfg/chainhash/v1.0.3
- chaincfg/v3.1.0
- connmgr/v3.1.0
- dcrutil/v4.0.0
- gcs/v3.0.0
- hdkeychain/v3.0.1
- rpc/jsonrpc/types/v3.0.0
- wire/v1.5.0

blockchain/v4.0.0 is not yet tagged, so it is updated to latest.

The non-release versions of dcrec, blake256, and certgen contain only
comment typo fixes and other changes that have no actual effect.
As such, this goes back to the latest tagged versions of these modules.

github.com/decred/slog is also updated from v1.1.0 to v1.2.0

Actual script versions are used when they are available. When they are
not available, a scriptVersionAssumed const is used.

gRPC: script class translation, docs update, and minor bump

The script classes from stdscript are now translated by the
ValidateAddress and DecodeTransation handlers to keep the responses
stable since the new stdscript types are now more fine-grained.

The tgen and tadd script types are added to both script type enums.

The gRPC server version is bumped from 7.12.0 to 7.13.0.

An example gRPC client app is added in rpc/grpc_example.