Follow DCP0010 once activated
This is currently being done using an extra getblockchaininfo RPC call
any time votes may need to be created.  While not ideal, and caching
of previously observed active and lockedin statuses could be utilized,
this is the bare minimum needed to make wallet follow the DCP0010 hard
fork if it is approved.