Use DCP0010 activation for VSP fee calculations
The VSP fee is a proportion of the subsidy received due to voting.
This value should change if/when DCP0010 activates so VSPs continue to
take the proper percentage of their fees.

In SPV mode, for the purposes of UTXO selection when needing to create
an extra output to pay a VSP fee, DCP0010 is assumed to have
activated.  Activation of the consensus rule is not possible to
determine in SPV mode, and assuming activation results in a large
enough UTXO to be created regardless if DCP0010 has activated or not.

Old stakepoold vestige code will continue to use the original subsidy
split rules.  Nobody should be purchasing stakepoold tickets anymore,
and this code should be removed when possible.