Fix compatibility with latest development dcrd
dcrd's JSON-RPC version received a major version bump for the removal
of several RPCs related to querying expired and missed tickets.  With
the activation of DCP0009 and the revocations of all older unspent
tickets, these methods are no longer needed by the wallet.

The "revoked" status for getstakeinfo is no longer used, and instead,
a status of "expired" or "missed" is used (which also implies that the
ticket has been revoked).  For this subtle breaking change,
dcrwallet's JSON-RPC API version also receives a major bump.