rpcserver: Allow signrawtransaction private keys.
When the wallet was updated to use stdaddr it did not account for the
fact that p2pk addresses no longer automatically convert their
stringized output to p2pkh addresses, so only the p2pk address ends up
in the map while utxos are often p2pkh address.

As a result, the signing code attempts to look up the key based on the
p2pkh address which is not in the map and errors despite the private key
being provided.

This resolves the issues by additionally adding the p2pkh variant to the
keys map for address that support it.