multi: Update dcrd pkgs to wip major versions
This commit updates the dcrd packages that have had a bump in their
major version to their current (non-released) wip versions. The
following modules/packages were updated:

blockchain/stake => v4 to v5
blockchain => v4 to v5
gcs => v3 to v4
rpc/jsonrpc/types => v3 to v4
rpcclient => v7 to v8

The only significant change API-wise that affected dcrwallet was the
removal of the extra arguments to stake.DetermineTxType and
stake.IsRRTx/IsSSGen. The determination of whether to check for the
additional rules pertaining to treasury and auto revocation agendas is
done by the transaction version field of the transaction itself, so that
eliminates the need for the additional flags. The official wallet
software never produced transactions that did not follow this standard,
therefore this is safe to do.

The replace directives added to go.mod are needed due to the fact that
some of the packages in dcrd (gcs/v4 and rpcclient/v8) depend on
blockchain/stake/v5 and their own respective go.mod are set to require
stake v5.0.0 (instead of a pre-release version).