Do not require additional newline prompting a hex seed
We currently require an extra blank line following the seed during
creation of a restored wallet.  This is due to two factors:

* During creation of a new wallet, we print the seed in PGP word list
  encoding using multiple lines, and users copy and paste this as is,
  newlines included.  This means our prompt must support reading
  multiple lines of input.

* Because the seed length is not required to be exactly 32 bytes (33
  words), we don't know when to stop prompting for seed words that are
  split across multiple lines.  An extra blank line after the seed is
  used to stop the prompting.

This prompting behavior is only needed for seeds encoded with the PGP
word list, and not seeds encoded in hex.  If the first line of input
is valid hex encoding, stop prompting immediately and do not require
the additional blank line following the seed.