Avoid needless balance calculations from ticket autobuyer
When the ticketbuyer.balancetomaintainabsolute flag is unset/zero,
there is no reason that the account balance needs to be calculated.
On busier wallets, avoiding this results in a significant performance
improvement.

This flag is already broken for a different reason (change can land in
other accounts, causing the account balance to drop below this
setting) but there are still some people using the flag despite this,
and compensating by setting it higher than they need.