Implement faster random UTXO lookup
This uses cursors seeking to a random key to randomly pick from UTXOs
in the bbolt db.  Previously, the randomization was implemented by
reading all UTXOs into memory and shuffling.

To avoid endless spinning, random attempts are only performed for half
of the total number of elements in the bucket, after which the
algorithm reverts back to reading everything again to shuffle the
remaining results.  For this to be possible, a new method is added to
the database interface returning the number of key/value pairs in a
bucket.

While here, remove the unused UnspentOutputsForAmount method.  We are
already in a new major version that has not been tagged yet.