vsp: Hard-code background context for async code.
The VSP client code was not handling contexts correctly. It was receiving
short-lived contexts bound to the lifespan of a single RPC request, and using
them to create feePayment instances. Most of the code in feePayment executes
async, so should be running on a long-lived context such as context.Background().