wallet: Prune chain instead of rooted tree
This modifies the pruning procedure of the sidechain forest to prune not
just entire trees, but instead to prune partial chains from trees.

This enables the ability to maintain headers in the forest for which its
ancestor chain has been included in the wallet's main chain, while the
successor chain has not yet been fully processed (for example, due to
missing cfilters).

While this does not make a functional difference at the moment, in the
future this will allow the SPV syncer to process chains asynchronously.