wallet: Avoid recalculating BlockHash
This changes the signature of the ExtendMainChain functions in the
wallet and udb to directly take header's block hash instead of
recalculating.

This avoids having to re-hash a header that has already been hashed
earlier in the call stack, reducing both cpu and allocated space
requirements during the initial sync process.